modified line 192

+++ test/Conversion/TorchToArith/basic.mlir
@@ -189,7 +189,7 @@ func.func @torch.aten.div.float(%arg0: !torch.float, %arg1: !torch.float) -> !to
 // CHECK:          %[[OUT:.*]] = torch_c.from_f64 %[[SUB:.*]]
 // CHECK:          return %[[OUT:.*]] : !torch.float
 func.func @torch.aten.div.int(%arg0: !torch.int, %arg1: !torch.int) -> !torch.float {
-  %0 = torch.aten.div.int %arg0, %arg1 : !torch.int, !torch.int -> !torch.float
+  %0 = torch.aten.div.int %arg0, %arg1 : !torch.float, !torch.float -> !torch.float
   return %0 : !torch.float
 }
